Last update. Implemented manual mode for EE robot in which we can control angles individually. Thonny code for ECME robot that does the same.

+++ ECME_robot/Pico_C_Code/RoboticArm_ECME/Inc/SPI.h
@@ -14,7 +14,7 @@
 static inline void cs_select();
 static inline void cs_deselect();
 void SPI_initialization();
-float MCP3208(int chan);
+int MCP3208(int chan);
 void read_encoders();
 
 #endif
